Print pointers in correct way.

+++ source/Tfe/Tfe.cpp
@@ -504,7 +504,7 @@ int tfe_activate_i(void)
 
 #ifdef TFE_DEBUG_INIT
     if(g_fh) fprintf(g_fh, "tfe_activate: Allocated memory successfully.\n");
-    if(g_fh) fprintf(g_fh, "\ttfe at $%08X, tfe_packetpage at $%08X\n", tfe, tfe_packetpage );
+    if(g_fh) fprintf(g_fh, "\ttfe at $%08X, tfe_packetpage at $%08X\n", uintptr_t(tfe), uintptr_t(tfe_packetpage) );
 #endif
 
 #ifdef DOS_TFE
